+++ src/routers/authRouter.ts
@@ -1,5 +1,5 @@
 import express, { Request, Response, Router } from 'express';
-import { getRequestCookie, removeRequestCookie } from '../util/cookieUtils';
+import { getRequestCookie, removeRequestCookie, setResponseCookie } from '../util/cookieUtils';
 import { generateCryptoUuid, isValidUuid } from '../util/tokenGenerator';
 import { dbPool } from '../db/db';
 import { ResultSetHeader, RowDataPacket } from 'mysql2/promise';
@@ -74,7 +74,7 @@ authRouter.get('/session', async (req: Request, res: Response) => {
       const newAuthSessionId: string = generateCryptoUuid();
       const newExpiryTimestamp: number = authSessionDetails.expiry_timestamp + dayMilliseconds;
 
-      await dbPool.execute(
+      const [resultSetHeader] = await dbPool.execute<ResultSetHeader>(
         `UPDATE
           auth_sessions
         SET
@@ -85,6 +85,9 @@ authRouter.get('/session', async (req: Request, res: Response) => {
           session_id = ?;`,
         [newAuthSessionId, newExpiryTimestamp, authSessionId]
       );
+
+      resultSetHeader.affectedRows > 0 &&
+        setResponseCookie(res, 'authSessionId', newAuthSessionId, dayMilliseconds, true);
     }
 
     res.json({ isValidAuthSession: true });
@@ -97,6 +100,7 @@ authRouter.get('/session', async (req: Request, res: Response) => {
     }
 
     res.json({ isValidAuthSession: false });
+    await logUnexpectedError(req, err);
   }
 });
 
@@ -108,37 +112,20 @@ authRouter.delete('/session', async (req: Request, res: Response) => {
     return;
   }
 
-  if (!isValidUuid(authSessionId)) {
-    removeRequestCookie(res, 'authSessionId');
-    res.json({});
-
-    return;
-  }
-
   removeRequestCookie(res, 'authSessionId');
   res.json({});
 
   try {
-    const [resultSetHeader] = await dbPool.execute<ResultSetHeader>(
-      `DELETE FROM
-        auth_sessions
-      WHERE
-        session_id = ?;`,
-      [authSessionId]
-    );
-
-    if (resultSetHeader.affectedRows === 0) {
-      await logUnexpectedError(req, null, 'Failed to delete auth_sessions.');
-    }
+    isValidUuid(authSessionId) &&
+      (await dbPool.execute<ResultSetHeader>(
+        `DELETE FROM
+          auth_sessions
+        WHERE
+          session_id = ?;`,
+        [authSessionId]
+      ));
   } catch (err: unknown) {
     console.log(err);
-
-    if (res.headersSent) {
-      await logUnexpectedError(req, err, 'Attempted to send two responses.');
-      return;
-    }
-
-    res.status(500).json({ message: 'Internal server error.' });
     await logUnexpectedError(req, err);
   }
 });